dev: 상품 컨트롤러 작성

+++ src/main/java/kr/hhplus/be/server/domain/product/controller/apidto/ProductSummary.java
@@ -1,3 +1,5 @@
 package kr.hhplus.be.server.domain.product.controller.apidto;
 
-public record ProductSummary(Long productId, String name, double price) {}
+import java.math.BigDecimal;
+
+public record ProductSummary(Long productId, String name, BigDecimal price) {}
